fix: relaxed constraints in `CloseOrUpdate` to allow composition.
It originally required the `prevInput` field to be `Nothing` for
every datum but there was no real security benefit to this and
it blocked composition with `Swap`. It also required updates
to be output to the same address but, since the owner directly
approved, this check was also unnecessary and interferred with
composition.