make it build with ghc-9.x (#14)
In order to make the off-chain code build with newer versions of
ghc, the code needed to migrate away from plutus-apps since it is
no longer being updated. Instead, the off-chain code now depends
on the cardano-node-emulator, cardano-api, and plutus github
repos. 

These newer dependencies require blst to be installed so the
installation instructions were edited to include it (the
instructions were take from the cardano-node install
instructions). The changes to the install instructions also
include usage of the new intersectmbo github repos as well as
some formatting changes.

Migrating to the new dependencies required a significant refactor
of the off-chain code (especially the test suite). While
refactoring, the CardanoSwaps haskell library (the src directory)
was re-organized to make supporting future versions of swaps
easier. The previous, top-level, mono-module would be annoying to
maintain so it was dropped in favor of just qualifying imports
by version numbers (similar to what the plutus repo does).

Building was tested with ghc-9.2.8, ghc-9.4.8, and ghc-9.6.4.
9.2.8 and 9.6.4 were successfull, but 9.4.8 failed. I am not sure
why; cabal threw an error when building ouroboros-consensus. I
did not seriously try to fix this since the other two versions
worked. 

The default ghc version was set to 9.6.4 but can be easily
changed by updating the cabal.project file's `with-compiler`
setting.